Remove support for skipping comments
The first step towards removing all Mode operations in the scanner.
They incur additional complexity and I don't want to use the feature
or maintain it.

This will make it harder to inherit upstream improvements to the
scanner, but we're probably going to have to modify it too much to get
any benefit there anyway.