Always scan all tokens
The original "text/scanner" package is intended to be used in many
circumstances parsing languages that are similar to go.  The consumer
may or may not care about certain types of tokens.  That's fine and
probably useful in certain cases, but it's more general and complex
than Golog needs.  It seems best to eliminate that code and the
associated maintenance/testing costs.