Linked list backed by a lexeme channel
The parser will eventually have to proceed arbitrarily forward through
available lexemes before it knows whether it's parsed a legitimate
term or not.  If not, it must backtrack through lexemes it's already
seen to try finding a term in a different way.

By parsing terms from a linked list, we get free backtracking (by
holding a pointer to an earlier point in the list) and the garbage
collector frees memory when we're certain not to need a specific
lexeme anymore.

There are almost certainly solutions to this problem which are more
time/space efficient, but we're not yet to the point that we care.