Convert parser to iteration (from channels)
Reading certain Prolog terms (op/3, use_module/2) can change the
parser's behavior by modifying the operator table.  With a naive
channel-based parser, it's possible for the parser to try reading
another term before the operator table has been modified.  This could
lead to a parse error.

The problem could be fixed by moving Prolog smarts into the parser.
However, it has to know *a lot* about Prolog.  Handling op/3 is easy,
but handling use_module/2 requires that the parser knows about modules
and their contents.  The Prolog execution environment already knows
about that, so it's silly to put it in the parser.

All of that implies that the TermReader client must be the one in
charge.  The reader should only try to read when the client asks it
to.  That way, the client has a chance to change the operator table if
necessary.

The lexer doesn't face this problem because Prolog doesn't allow the
dynamic code to change what the lexer sees as tokens.