Rename scanner to lex
Best practices for Go require typing and reading package names all
over the place.  Both scanner and lexer are acceptable terms for this
component.  Because of the software tool lex, that abbreviation is
well known.  Before starting on this project, I had actually never
heard the term scanner, so lex would have been more familiar to me as
a newbie.

All these reasons justify the change in my mind.  As icing on the
cake, it let me do the cute type name lex.Eme for a token :-)