Rename variables after reading terms
Variables are scoped to a single term.  Developers are free to reuse
variable names in other terms and they refer to completely different
variables.  To accomplish this behavior, Golog now renames variables
to a unique, internal representation after reading a term.

This corrected behavior was required by TestAppend which intentionally
uses the variable name A in two separate clauses.  Without variable
renaming, the variables in each clause are falsely thought to refer to
the same value.

This also required some work on Bindings so that they can retain a
human-readable variable mapping.  This makes it possible to do

    proofs := m.ProveAll(`foo(X).`)
    x := proofs[0].ByName_("X")

without the caller creating variables manually.