Make recursive clauses work
This adds a test for append/3 which triggers the second clause which
calls itself.  My old code failed because the callees variables
retained the values of the caller.  The former should have started out
as fresh variables.

I addressed the problem by renaming all variables in a clause before
attempting unification.  I saw this same approach in ioctl's
JavaScript implementation[1].  There are probably much more efficient
ways to accomplish this, but this way works.  I can optimize later, if
it proves necessary.

1: http://ioctl.org/logic/jsprolog-history