Rudimentary Golog top-level
The best way to run the top-level is with ./golog.sh  That script
compiles bin/golog.go into a binary and launches it wrapped with
rlwrap (to support line editing and history).  Passing a Prolog source
file as the only command line argument loads that code into the
database.

For now, the top-level displays all solutions provable from a query.
There's no way to exit early.