Support double quoted strings as code lists
For now they're desugared to a chain of cons cells containing
integers.  Eventually I want to provide a CodeList value which
implements the Term interface.  I suspect this will offer better speed
and memory performance.  As long as it behaves like a list from
Prolog, it can be whatever I want underneath.