Completely rewrite Golog machine
The previous machine was so complex that I couldn't understand what it
was doing, even though I wrote it during the last two weeks.  The new
machine is conceptually simpler and should have a much better chance
of correctly handling cuts that cut across multiple cut parents.

This version is also 80 lines shorter.