Implement atom_codes/2
This implements the ISO predicate defined in §8.16.5  It also
implements new infrastructure for writing tests in pure Prolog.
Include a Prolog file in t/ to execute every predicate in that file as
a test.  Tests can be marked as expected to fail; otherwise, they're
assumed to succeed if the predicate succeeds.

This should make it much easier to test new predicates.  Without this,
I have to hand-write Go tests which are tedious and error prone
because of all the boilerplate.