Fix a bug in findall/3
findall/3 was searching for solutions in the goal it was given along
with the goals in its own continuation.  This was the cause of an
earlier findall/3 test failure (which I didn't solve at the time).

The fix is pretty obvious: before finding all solutions to the given
goal, clear the conjunctions stack so that the continuation can't
affect the solutions.  Since this is done in a "forked" Golog machine,
it has no effect on the original machine that called findall/3.  This
makes me love the immutability approach I've taken.