Optimization for cut barriers
A cut barrier is a reference point that cuts can use to eliminate
disjunctions.  It's redundant to have multiple cut barriers in a row
because cut to any one of them has the same effect: removing all
disjunctions in front of them.  This commit doesn't add another cut
barrier if there's one already on top of the stack.

I encountered this problem while debugging some code that generated a
lot of cut barriers.  The machine had dozens of redundant cut barriers
filling up memory and making it hard to debug the machine.  This new
code makes the machine dump easier to read.

Performance was not a concern here and I did no profiling.