Benchmark for naive reverse
This is apparently the traditional benchmark for testing a Prolog
implementations performance.  Here are results for all benchmarks as
of this commit:

    True               46715 ns/op
    Append            353850 ns/op
    Maplist           828163 ns/op
    NaiveReverse    94176150 ns/op

The reverse benchmark implies 5,230 LIPS.  That's not too bad
considering I've done no work on performance, never implemented Prolog
before and have no database indexing.  It'll be a fun weekend project
to improve the performance.