Add a benchmark similar to some DCG rules
In this case, there is a single predicate which describes a person's
name on the front of a list.  DCGs often have a format similar to
this.  This also tests predicates with many clauses of which only a
single clause matches.

This benchmark is twice as fast with term indexing as it is without.
Earlier benchmarks didn't show much improvement from indexing.  That's
because the predicates in those benchmarks only have 1 or 2 clauses.