Run TAP tests that don't have names
library(tap) recently introduced the ability to write tests that have
no name.  The test is described as a term which represents both the
test name and the test logic.  Support those kinds of tests in the
Golog test suite.

This requires us to recognize when library(tap) is imported so that
only those terms after the import are considered as tests.