Clear disjunctions before seeking \+ goal
If the \+ goal is sought with disjunctions in place, a failing goal
causes Golog to backtrack to the other disjunctions.  However, those
disjunctions are not relevant to the goal at hand.  They're only
relevant higher up the call stack.

The included test case failed before this fix.  It passes now.
SWI-Prolog passes the test case too.